Backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.40.0
motor==3.7.1
pymongo==4.14.1
langgraph==0.6.7
openai==1.107.3
python-dotenv==1.1.1

